refactor: restructure service layer with domain-based separation (#10)
- Split large services into Query and Management services
- Reorganize service packages by domain (employee/, attendance/, etc.)
- Update controllers to inject specialized services
- Apply Single Responsibility Principle and interface segregation
- Maintain backward compatibility with existing REST APIs
