+++ src/main/java/com/furkanerd/hr_management_system/controller/SalaryController.java
@@ -6,7 +6,7 @@ import com.furkanerd.hr_management_system.model.dto.response.ApiResponse;
 import com.furkanerd.hr_management_system.model.dto.response.PaginatedResponse;
 import com.furkanerd.hr_management_system.model.dto.response.salary.ListSalaryResponse;
 import com.furkanerd.hr_management_system.model.dto.response.salary.SalaryDetailResponse;
-import com.furkanerd.hr_management_system.service.SalaryService;
+import com.furkanerd.hr_management_system.service.salary.SalaryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
