+++ src/main/java/com/furkanerd/hr_management_system/service/attendance/impl/AttendanceManagementServiceImpl.java
@@ -1,79 +1,46 @@
-package com.furkanerd.hr_management_system.service.impl;
+package com.furkanerd.hr_management_system.service.attendance.impl;
 
-import com.furkanerd.hr_management_system.constants.SortFieldConstants;
 import com.furkanerd.hr_management_system.exception.*;
 import com.furkanerd.hr_management_system.mapper.AttendanceMapper;
 import com.furkanerd.hr_management_system.model.dto.request.attendance.AttendanceCreateRequest;
-import com.furkanerd.hr_management_system.model.dto.request.attendance.AttendanceFilterRequest;
 import com.furkanerd.hr_management_system.model.dto.request.attendance.AttendanceUpdateRequest;
-import com.furkanerd.hr_management_system.model.dto.response.PaginatedResponse;
 import com.furkanerd.hr_management_system.model.dto.response.attendance.AttendanceDetailResponse;
-import com.furkanerd.hr_management_system.model.dto.response.attendance.ListAttendanceResponse;
 import com.furkanerd.hr_management_system.model.entity.Attendance;
 import com.furkanerd.hr_management_system.model.entity.Employee;
 import com.furkanerd.hr_management_system.repository.AttendanceRepository;
 import com.furkanerd.hr_management_system.repository.EmployeeRepository;
-import com.furkanerd.hr_management_system.service.AttendanceService;
-import com.furkanerd.hr_management_system.specification.AttendanceSpecification;
-import com.furkanerd.hr_management_system.util.PaginationUtils;
-import com.furkanerd.hr_management_system.util.SortFieldValidator;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
+import com.furkanerd.hr_management_system.service.attendance.AttendanceManagementService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 import java.util.UUID;
 
+
 @Service
-class AttendanceServiceImpl implements AttendanceService {
+@Transactional
+class AttendanceManagementServiceImpl implements AttendanceManagementService {
 
     private final AttendanceRepository attendanceRepository;
     private final EmployeeRepository employeeRepository;
     private final AttendanceMapper attendanceMapper;
 
-    public AttendanceServiceImpl(AttendanceRepository attendanceRepository, EmployeeRepository employeeRepository, AttendanceMapper attendanceMapper) {
+    private static final Duration MIN_WORK_DURATION = Duration.ofHours(8);
+    private static final LocalTime CHECK_IN_START_TIME = LocalTime.of(6, 0);
+    private static final LocalTime CHECK_IN_END_TIME = LocalTime.of(10, 0);
+
+
+    public AttendanceManagementServiceImpl(AttendanceRepository attendanceRepository, EmployeeRepository employeeRepository, AttendanceMapper attendanceMapper) {
         this.attendanceRepository = attendanceRepository;
         this.employeeRepository = employeeRepository;
         this.attendanceMapper = attendanceMapper;
     }
 
-    private static final Duration MIN_WORK_DURATION = Duration.ofHours(8);
-    private static final LocalTime CHECK_IN_START_TIME = LocalTime.of(6, 0);
-    private static final LocalTime CHECK_IN_END_TIME = LocalTime.of(10, 0);
-
     @Override
-    public PaginatedResponse<ListAttendanceResponse> listAllAttendance(int page, int size, String sortBy, String sortDirection, AttendanceFilterRequest filterRequest) {
-        String validatedSortBy = SortFieldValidator.validate(SortFieldConstants.ATTENDANCE_SORT_FIELD, sortBy);
-        Pageable pageable = PaginationUtils.buildPageable(page, size, validatedSortBy, sortDirection);
-
-        Specification<Attendance> specification = AttendanceSpecification.withFilters(filterRequest);
-
-        Page<Attendance> attendancePage = attendanceRepository.findAll(specification, pageable);
-        List<ListAttendanceResponse> responseList = attendanceMapper.attendancesToListAttendanceResponse(attendancePage.getContent());
-        return PaginatedResponse.of(
-                responseList,
-                attendancePage.getTotalElements(),
-                page,
-                size
-        );
-    }
-
-    @Override
-    public AttendanceDetailResponse getAttendanceById(UUID id) {
-        return attendanceMapper.attendanceToAttendanceDetailResponse(attendanceRepository.findById(id)
-                .orElseThrow(() -> new AttendanceNotFoundException("Attendance with id: " + id))
-        );
-    }
-
-    @Override
-    @Transactional
     public AttendanceDetailResponse createAttendance(AttendanceCreateRequest createRequest) {
-        Employee employee =employeeRepository.findById(createRequest.employeeId())
+        Employee employee = employeeRepository.findById(createRequest.employeeId())
                 .orElseThrow(() -> new EmployeeNotFoundException(createRequest.employeeId()));
 
         if (attendanceRepository.existsByEmployeeIdAndDate(createRequest.employeeId(), createRequest.date())) {
@@ -94,11 +61,9 @@ class AttendanceServiceImpl implements AttendanceService {
                 .build();
 
         return attendanceMapper.attendanceToAttendanceDetailResponse(attendanceRepository.save(attendance));
-
     }
 
     @Override
-    @Transactional
     public AttendanceDetailResponse autoCheckIn(String employeeEmail) {
         Employee employee = employeeRepository.findByEmail(employeeEmail)
                 .orElseThrow(() -> new EmployeeNotFoundException(employeeEmail));
@@ -123,9 +88,7 @@ class AttendanceServiceImpl implements AttendanceService {
         return attendanceMapper.attendanceToAttendanceDetailResponse(attendanceRepository.save(attendance));
     }
 
-
     @Override
-    @Transactional
     public AttendanceDetailResponse autoCheckOut(String employeeEmail) {
         Employee employee = employeeRepository.findByEmail(employeeEmail)
                 .orElseThrow(() -> new EmployeeNotFoundException(employeeEmail));
@@ -143,13 +106,10 @@ class AttendanceServiceImpl implements AttendanceService {
 
         attendance.setCheckOutTime(now);
         return attendanceMapper.attendanceToAttendanceDetailResponse(attendanceRepository.save(attendance));
-
     }
 
     @Override
-    @Transactional
     public AttendanceDetailResponse updateAttendance(UUID id, AttendanceUpdateRequest updateRequest) {
-
         Attendance toUpdateAttendance = attendanceRepository.findById(id)
                 .orElseThrow(() -> new AttendanceNotFoundException("Attendance with id: " + id));
 
@@ -168,63 +128,12 @@ class AttendanceServiceImpl implements AttendanceService {
         return attendanceMapper.attendanceToAttendanceDetailResponse(attendanceRepository.save(toUpdateAttendance));
     }
 
-    @Override
-    public PaginatedResponse<ListAttendanceResponse> getAttendanceByEmployee(String employeeEmail, int page, int size, String sortBy, String sortDirection, AttendanceFilterRequest filterRequest) {
-        Employee employee =  employeeRepository.findByEmail(employeeEmail)
-                .orElseThrow(() -> new EmployeeNotFoundException(employeeEmail));
-
-        String validatedSortBy = SortFieldValidator.validate(SortFieldConstants.ATTENDANCE_SORT_FIELD, sortBy);
-        Pageable pageable = PaginationUtils.buildPageable(page, size, validatedSortBy, sortDirection);
-
-        Specification<Attendance> baseSpec = AttendanceSpecification.withFilters(filterRequest);
-
-        Specification<Attendance> specification = (baseSpec != null)
-                ? baseSpec.and((root, query, cb) -> cb.equal(root.get(SortFieldConstants.EMPLOYEE_SORT_FIELD).get("id"), employee.getId()))
-                : (root, query, cb) -> cb.equal(root.get(SortFieldConstants.EMPLOYEE_SORT_FIELD).get("id"), employee.getId());
-
-        Page<Attendance> attendancePage = attendanceRepository.findAll(specification, pageable);
-        List<ListAttendanceResponse> responseList = attendanceMapper.attendancesToListAttendanceResponse(attendancePage.getContent());
-
-        return PaginatedResponse.of(
-                responseList,
-                attendancePage.getTotalElements(),
-                page,
-                size
-        );
-    }
 
     @Override
     public void deleteAttendance(UUID id) {
         Attendance attendance = attendanceRepository.findById(id)
                 .orElseThrow(() -> new AttendanceNotFoundException("No attendance found for id " + id));
         attendanceRepository.delete(attendance);
-    }
-
-    @Override
-    public PaginatedResponse<ListAttendanceResponse> getAllAttendanceByEmployee(UUID id, int page, int size, String sortBy, String sortDirection, AttendanceFilterRequest filterRequest) {
-        boolean exists = employeeRepository.existsById(id);
-        if (!exists) {
-            throw new EmployeeNotFoundException(id);
-        }
-
-        String validatedSortBy = SortFieldValidator.validate(SortFieldConstants.ATTENDANCE_SORT_FIELD, sortBy);
-        Pageable pageable = PaginationUtils.buildPageable(page, size, validatedSortBy, sortDirection);
-
-        Specification<Attendance> baseSpec = AttendanceSpecification.withFilters(filterRequest);
-
-        Specification<Attendance> specification = (baseSpec != null)
-                ? baseSpec.and((root, query, cb) -> cb.equal(root.get("employee").get("id"), id))
-                : (root, query, cb) -> cb.equal(root.get("employee").get("id"), id);
-
-        Page<Attendance> attendancePage = attendanceRepository.findAll(specification, pageable);
-        List<ListAttendanceResponse> responseList = attendanceMapper.attendancesToListAttendanceResponse(attendancePage.getContent());
-
-        return PaginatedResponse.of(
-                responseList,
-                attendancePage.getTotalElements(),
-                page,
-                size
-        );
     }
 
     /**
@@ -248,18 +157,6 @@ class AttendanceServiceImpl implements AttendanceService {
         }
     }
 
-    /**
-     * Checks: Has at least the minimum working time passed between check-in and check-out times?
-     *
-     * @param checkInTime  Check-in time
-     * @param checkOutTime Check-out time
-     * @return True if the minimum working time has passed, false otherwise
-     */
-    private boolean isMinimumWorkDurationMet(LocalTime checkInTime, LocalTime checkOutTime) {
-        Duration duration = Duration.between(checkInTime, checkOutTime);
-        return duration.compareTo(MIN_WORK_DURATION) >= 0;
-    }
-
 
     /**
      * Checks if the check-out time is valid
@@ -274,5 +171,17 @@ class AttendanceServiceImpl implements AttendanceService {
         if (!isMinimumWorkDurationMet(checkInTime, checkOutTime)) {
             throw new InvalidAttendanceTimeException("Check-out time must be at least 8 hours after check-in time");
         }
+    }
+
+    /**
+     * Checks: Has at least the minimum working time passed between check-in and check-out times?
+     *
+     * @param checkInTime  Check-in time
+     * @param checkOutTime Check-out time
+     * @return True if the minimum working time has passed, false otherwise
+     */
+    private boolean isMinimumWorkDurationMet(LocalTime checkInTime, LocalTime checkOutTime) {
+        Duration duration = Duration.between(checkInTime, checkOutTime);
+        return duration.compareTo(MIN_WORK_DURATION) >= 0;
     }
 }
